🚚 refactor: extract agent sandbox package
