+++ internal/agent/coretools_builder.go
@@ -2,12 +2,13 @@ package agent
 
 import (
 	"github.com/CherryHQ/stella/internal/agent/coretools"
+	"github.com/CherryHQ/stella/internal/agent/sandbox"
 	"github.com/CherryHQ/stella/pkg/tools"
 	plugintools "github.com/CherryHQ/stella/plugins/tools"
 )
 
 // buildSandboxCoreTools creates core tools using the active sandbox session.
-func buildSandboxCoreTools(session *runnerSession, bc plugintools.BuildContext) []tools.Tool {
+func buildSandboxCoreTools(session *sandbox.Session, bc plugintools.BuildContext) []tools.Tool {
 	if session == nil || session.Session() == nil {
 		return nil
 	}
