+++ internal/agent/coretools_builder_test.go
@@ -4,37 +4,35 @@ import (
 	"context"
 	"testing"
 
+	"github.com/CherryHQ/stella/internal/agent/sandbox"
 	pkgplugins "github.com/CherryHQ/stella/pkg/plugins"
-	"github.com/CherryHQ/stella/pkg/sandbox"
+	pkgsandbox "github.com/CherryHQ/stella/pkg/sandbox"
 	plugintools "github.com/CherryHQ/stella/plugins/tools"
 )
 
-func fakeRunnerSession(_ string, sess sandbox.Session) *runnerSession {
-	return &runnerSession{
-		session: sess,
-		policy:  sandbox.Policy{},
-	}
+func fakeRunnerSession(_ string, sess pkgsandbox.Session) *sandbox.Session {
+	return sandbox.NewSession(sess)
 }
 
 type fakeSession struct {
 	alive  bool
-	policy sandbox.Policy
+	policy pkgsandbox.Policy
 }
 
-func (f *fakeSession) Policy() sandbox.Policy { return f.policy }
-func (f *fakeSession) Close() error           { return nil }
-func (f *fakeSession) Alive() bool            { return f.alive }
+func (f *fakeSession) Policy() pkgsandbox.Policy { return f.policy }
+func (f *fakeSession) Close() error              { return nil }
+func (f *fakeSession) Alive() bool               { return f.alive }
 func (f *fakeSession) Done() <-chan struct{} {
 	ch := make(chan struct{})
 	close(ch)
 	return ch
 }
 
-func (f *fakeSession) Exec(_ context.Context, _ string, _ sandbox.ExecOptions) (sandbox.ExecResult, error) {
-	return sandbox.ExecResult{Stdout: "ok", ExitCode: 0}, nil
+func (f *fakeSession) Exec(_ context.Context, _ string, _ pkgsandbox.ExecOptions) (pkgsandbox.ExecResult, error) {
+	return pkgsandbox.ExecResult{Stdout: "ok", ExitCode: 0}, nil
 }
 
-func (f *fakeSession) StartProcess(_ context.Context, _ sandbox.ProcessRequest) (sandbox.ProcessHandle, error) {
+func (f *fakeSession) StartProcess(_ context.Context, _ pkgsandbox.ProcessRequest) (pkgsandbox.ProcessHandle, error) {
 	return nil, nil
 }
 func (f *fakeSession) ResolvePath(path string) (string, error) { return path, nil }
