+++ internal/agent/paths.go
@@ -1,10 +1,9 @@
 package agent
 
 import (
-	"fmt"
 	"path/filepath"
 
-	"github.com/CherryHQ/stella/internal/config"
+	"github.com/CherryHQ/stella/internal/agent/sandbox"
 )
 
 // runnerPaths keeps only the runner's primary inputs.
@@ -16,56 +15,20 @@ type runnerPaths struct {
 	ProjectRoot string
 }
 
-// sandboxPaths is the minimal path set sandbox policy creation depends on.
-// Sandbox execution is defined entirely by the user-scoped writable root and an
-// internal working directory derived from that root.
-type sandboxPaths struct {
-	StellaHome  string
-	UserRoot    string
-	WorkDir     string
-	AgentRoot   string
-	ProjectRoot string
-}
-
 // resolveRunnerPaths converts GoRunnerConfig into the minimal path set the
 // runner actually owns. Derived paths stay as methods/helpers instead of being
 // stored as extra fields.
 func resolveRunnerPaths(cfg GoRunnerConfig) runnerPaths {
-	paths, _ := resolveSandboxPaths(cfg)
+	paths, _ := sandbox.ResolvePaths(sandbox.Config{
+		StellaHome: cfg.StellaHome,
+		UserRoot:   cfg.UserRoot,
+	})
 	return runnerPaths{
 		StellaHome:  paths.StellaHome,
 		AgentRoot:   cfg.AgentRoot,
 		UserRoot:    paths.UserRoot,
 		ProjectRoot: cfg.ProjectRoot,
 	}
-}
-
-func resolveSandboxWorkingDir(cfg GoRunnerConfig, userRoot string) string {
-	return userRoot
-}
-
-func resolveSandboxPaths(cfg GoRunnerConfig) (sandboxPaths, error) {
-	stellaHome := cfg.StellaHome
-	if stellaHome == "" {
-		stellaHome = config.StellaHome()
-	}
-	if cfg.UserRoot == "" {
-		return sandboxPaths{}, fmt.Errorf("user_root is required")
-	}
-
-	userRoot, err := filepath.Abs(cfg.UserRoot)
-	if err != nil {
-		return sandboxPaths{}, fmt.Errorf("resolve user_root: %w", err)
-	}
-	workDir := resolveSandboxWorkingDir(cfg, userRoot)
-
-	return sandboxPaths{
-		StellaHome:  stellaHome,
-		UserRoot:    userRoot,
-		WorkDir:     workDir,
-		AgentRoot:   cfg.AgentRoot,
-		ProjectRoot: cfg.ProjectRoot,
-	}, nil
 }
 
 // resolveToolsBinDir always returns empty for docker: the container image has
@@ -80,18 +43,4 @@ func (p runnerPaths) stellaSkillsDir() string {
 
 func (p runnerPaths) stellaAgentsDir() string {
 	return filepath.Join(p.StellaHome, ".agents", "agents")
-}
-
-// sandboxProcessEnv builds the baseline process environment injected into
-// sandboxed docker commands. The docker container already provides its own
-// rootfs and image-baked HOME, so we leave HOME alone and let the image's user
-// home stand — that's what lets tools installed in the image (mise tree, shell
-// rc files, shims) remain reachable at runtime regardless of the workspace
-// bind-mount path.
-func sandboxProcessEnv(paths sandboxPaths) map[string]string {
-	env := map[string]string{}
-	if paths.StellaHome != "" {
-		env["STELLA_HOME"] = paths.StellaHome
-	}
-	return env
 }
