+++ internal/agent/sandbox/dood_test.go
@@ -1,4 +1,4 @@
-package agent
+package sandbox
 
 import (
 	"strings"
