+++ internal/agent/sandbox/trace.go
@@ -1,4 +1,4 @@
-package agent
+package sandbox
 
 import (
 	"fmt"
